refactor: 프록시 관련 로직을 별도 클래스로 분리
- domain/proxy/proxy.handler.ts로 추가
